A small scale e-commerce application
This is a small scale e-commerce application allowing merchants to sell unique items with different offers and customer can buy items easily. Customer can also maintain a cart for buying items. Rewards are awarded to merchants and customers on conditions defined by company. As on now, 5 merchants and 5 customers are predefined in the application.